taskmanager/auth-service/requirements.txt
asgiref==3.8.1
charset-normalizer==3.3.2
Django==4.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
django-redis==5.3.0
